update default data

+++ app/Modules/Role/Repositories/RoleRepository.ts
@@ -2,6 +2,7 @@ import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import { RoleValidationDto } from '../Validations/RoleValidation'
 import { AuthenticationException } from '@adonisjs/auth/build/standalone'
 import { EXCEPTION_CODE, EXCEPTION_MESSAGE } from 'App/Constants/String'
+import { ROLE_DEFAULT } from 'App/Constants/RolePermission'
 import Role from './Entities/Role'
 
 export default class RoleRepository {
@@ -61,8 +62,19 @@ export default class RoleRepository {
     return await this.find(role.id)
   }
 
+  private isInvalidRoleToModify(nameCheck: string): boolean {
+    const defaultRole: Array<string> = [];
+    for (const roleDefault of Object.values(ROLE_DEFAULT)) {
+      defaultRole.push(roleDefault)
+    }
+    return defaultRole.includes(nameCheck)
+  }
+
   public async update(id: number, data: RoleValidationDto) {
     const role = await this.find(id)
+    if (this.isInvalidRoleToModify(role.name)) {
+      throw new AuthenticationException(EXCEPTION_MESSAGE.E_INVALID_MODIFY_ROLE, EXCEPTION_CODE.E_INVALID_MODIFY_ROLE)
+    }
     role.name = data.name
     await role.save()
     await role.related('permissions').sync(data.permissions)
@@ -71,6 +83,9 @@ export default class RoleRepository {
 
   public async destroy(id: number) {
     const role = await this.find(id)
+    if (this.isInvalidRoleToModify(role.name)) {
+      throw new AuthenticationException(EXCEPTION_MESSAGE.E_INVALID_MODIFY_ROLE, EXCEPTION_CODE.E_INVALID_MODIFY_ROLE)
+    }
     await role.delete()
   }
 }
